+++ packages/LineChart/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.8.0](https://github.com/dpc-sdp/myvictoria-vic-gov-au/tree/master/packages/LineChart/compare/v1.7.0...v1.8.0) (2021-03-17)
+
+**Note:** Version bump only for package @dpc-sdp/myvic-line-chart
+
+
+
+
+
+
+# [1.7.0](https://github.com/dpc-sdp/myvictoria-vic-gov-au/tree/master/packages/LineChart/compare/v1.6.1...v1.7.0) (2021-02-12)
+
+**Note:** Version bump only for package @dpc-sdp/myvic-line-chart
+
+
+
+
+
+# [1.6.0](https://github.com/dpc-sdp/myvictoria-vic-gov-au/tree/master/packages/LineChart/compare/v1.5.3...v1.6.0) (2020-11-19)
+
+**Note:** Version bump only for package @dpc-sdp/myvic-line-chart
+
+
+
+
+
+## [1.5.3](https://github.com/dpc-sdp/myvictoria-vic-gov-au/tree/master/packages/LineChart/compare/v1.5.1...v1.5.3) (2020-11-18)
+
+**Note:** Version bump only for package @dpc-sdp/myvic-line-chart
+
+
+
+
+
 ## [1.5.2](https://github.com/dpc-sdp/myvictoria-vic-gov-au/tree/master/packages/LineChart/compare/v1.5.1...v1.5.2) (2020-10-27)
 
 **Note:** Version bump only for package @dpc-sdp/myvic-line-chart